Move out some code from lens-proxy.ts
